package renamed to usb_tree_lib, main.go renamed to reflect its usage, remaining lib wrapper functions completed.
