+++ app/app.go
@@ -2,7 +2,7 @@ package main
 
 import (
 	"context"
-	"toollib"
+	"usb_tree_lib"
 )
 
 // App struct
@@ -21,7 +21,7 @@ func (a *App) startup(ctx context.Context) {
 	a.ctx = ctx
 }
 
-// Wrapper for toollib.HelloWorld()
+// Wrapper for usb_tree_lib.HelloWorld()
 func (a *App) Greet() string {
-	return toollib.HelloWorld()
+	return usb_tree_lib.HelloWorld()
 }
