+++ cli/go.mod
@@ -2,6 +2,6 @@ module main
 
 go 1.25.0
 
-replace toollib => ../lib
+replace usb_tree_lib => ../lib
 
-require toollib v0.0.0-00010101000000-000000000000
+require usb_tree_lib v0.0.0-00010101000000-000000000000
