+++ cli/main.go
@@ -2,9 +2,9 @@ package main
 
 import (
 	"fmt"
-	"toollib"
+
 )
 
 func main() {
-	fmt.Printf("%v\n", toollib.HelloWorld())
+	fmt.Printf("")
 }
